requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
APScheduler==3.10.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
Django==4.2
django-adminlte3==0.1.6
environs==9.5.0
frozenlist==1.3.3
idna==3.4
loguru==0.6.0
magic-filter==1.0.9
marshmallow==3.19.0
multidict==6.0.4
packaging==23.0
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
redis==4.5.4
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
win32-setctime==1.1.0
yarl==1.8.2
django-import-export
openpyxl